関数上限に合わせて分割
